requirements.txt
openai==1.13.3
azure-identity==1.15.0
azure-functions==1.20.0
azure-core==1.30.2
azure-search-documents==11.4.0
azure-cosmos==4.7.0
python-dotenv==1.0.1
pyyaml==6.0.1
numpy==1.23.5
pandas==1.5.3
ijson